Initial Commit for the Real time chat
Small parts are pulled from https://github.com/ajgoldenwings/Asp-Net-Blazor-Forum-Square
This uses signalR. I do need to clean some of the front end and use more of MatBlazor and persist the forum messages as well. That is something I would like to do next and that would be the scope of the contribution I would like to do. Probably notifications as well afterwards.